website_analyzer/requirements.txt
# Dipendenze per il Website Analyzer AI
flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
beautifulsoup4==4.10.0
python-whois==0.7.3
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
